% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_preprocessing.R
\name{add_centralities}
\alias{add_centralities}
\title{Adds centrality metrics to graph}
\usage{
add_centralities(g, g0)
}
\arguments{
\item{g}{target graph to augment}

\item{g0}{master igraph with full graph directionality info}
}
\description{
Builds centrality metrics from inbound tie information
}
