% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_preprocessing.R
\name{build_edge_network}
\alias{build_edge_network}
\title{Produce pseudo-network containing all possible edges}
\usage{
build_edge_network(g)
}
\arguments{
\item{g}{master igraph}
}
\description{
This allows different edges to be dynamically hidden/masked, without reloading a different graph
}
