% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_preprocessing.R
\name{build_network_directions}
\alias{build_network_directions}
\title{Extract network directionalities from master igraph}
\usage{
build_network_directions(g, slug = "g")
}
\arguments{
\item{g}{master igraph}

\item{slug}{name slug for this graph}
}
\description{
Builds directed, mutual, and collapsed (any direction) networks from a complete master network
}
