% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{load_qualtrics_file}
\alias{load_qualtrics_file}
\title{Load qualtrics survey from zip or csv file}
\usage{
load_qualtrics_file(file_qualtrics)
}
\arguments{
\item{file_qualtrics}{list(datapath :: string path, type :: mime type)}
}
\description{
Load qualtrics survey from zip or csv file
}
