% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{qualtrics_to_igraph}
\alias{qualtrics_to_igraph}
\title{Convert survey dataframe into an igraph network for the rest of the pipeline}
\usage{
qualtrics_to_igraph(
  df.qualtrics,
  qidNetwork = "network",
  qidSudo = "pseudonym",
  qidConsent = "consent",
  ansAdvice = "advice",
  ansSupport = "support",
  ansConsent = "I agree"
)
}
\arguments{
\item{df.qualtrics}{dataframe derived from qualtrics file}

\item{qidNetwork}{QID for network roster question}

\item{qidSudo}{QID for pseudonym question}

\item{qidConsent}{QID for anonymity consent question}

\item{ansAdvice}{survey response for an advice network edge}

\item{ansSupport}{survey response for a support network edge}

\item{ansConsent}{survey response for anonymity consent}
}
\description{
Convert survey dataframe into an igraph network for the rest of the pipeline
}
