% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zkc.qualtrics.df}
\alias{zkc.qualtrics.df}
\title{Zachary Karate Club Network virtual Qualtrics survey response file}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 55 columns.
}
\usage{
zkc.qualtrics.df
}
\description{
Zachary Karate Club Network virtual Qualtrics survey response file
}
\keyword{datasets}
